requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.0.0
pdf2image>=1.17.0,<2
Pillow>=10.0.0
requests>=2.0.0
